// www/src/app/curriculo/models/skill.model.ts
export class Skill {
    id: number;
    developerId: number;
    aditionalInformation: string;
    linkCrud: string;
    ionic: number;
    reactJS: number;
    reactNative: number;
    android: number;
    flutter: number;
    sWIFT: number;
    iOS: number;
    hTML: number;
    cSS: number;
    bootstrap: number;
    jquery: number;
    angularJs: number;
    angular: number;
    java: number;
    python: number;
    flask: number;
    aspNetMVC: number;
    aspNetWebForm: number;
    c: number;
    csharp: number;
    nodeJS: number;
    expressNodeJs: number;
    cake: number;
    django: number;
    majento: number;
    pHP: number;
    vue: number;
    wordpress: number;
    ruby: number;
    mySQLServer: number;
    mySQL: number;
    salesforce: number;
    photoshop: number;
    illustrator: number;
    sEO: number;
    laravel: number;

    constructor() {
        this.id = 0;
        this.developerId = 0;
        this.ionic = 0;
        this.reactJS = 0;
        this.reactNative = 0;
        this.android = 0;
        this.flutter = 0;
        this.sWIFT = 0;
        this.iOS = 0;
        this.hTML = 0;
        this.cSS = 0;
        this.bootstrap = 0;
        this.jquery = 0;
        this.angularJs = 0;
        this.angular = 0;
        this.java = 0;
        this.python = 0;
        this.flask = 0;
        this.aspNetMVC = 0;
        this.aspNetWebForm = 0;
        this.c = 0;
        this.csharp = 0;
        this.nodeJS = 0;
        this.expressNodeJs = 0;
        this.cake = 0;
        this.django = 0;
        this.majento = 0;
        this.pHP = 0;
        this.vue = 0;
        this.wordpress = 0;
        this.ruby = 0;
        this.mySQLServer = 0;
        this.mySQL = 0;
        this.salesforce = 0;
        this.photoshop = 0;
        this.illustrator = 0;
        this.sEO = 0;
        this.laravel = 0;
    }
}