// www/src/app/curriculo/models/knowledge.model.ts
export class Knowledge {
    ionic: number;
    reactJS: number;
    reactNative: number;
    android: number;
    flutter: number;
    swift: number;
    ios: number;
    html: number;
    css: number;
    bootstrap: number;
    jquery: number;
    angularJs: number;
    angular: number;
    java: number;
    python: number;
    flask: number;
    aspNetMVC: number;
    aspNetWebForm: number;
    c: number;
    csharp: number;
    nodeJS: number;
    expressNodeJs: number;
    cake: number;
    django: number;
    majento: number;
    php: number;
    vue: number;
    wordpress: number;
    ruby: number;
    mySQLServer: number;
    mySQL: number;
    salesforce: number;
    photoshop: number;
    illustrator: number;
    seo: number;
    laravel: number;

    constructor() {
        this.ionic = 0;
        this.reactJS = 0;
        this.reactNative = 0;
        this.android = 0;
        this.flutter = 0;
        this.swift = 0;
        this.ios = 0;
        this.html = 0;
        this.css = 0;
        this.bootstrap = 0;
        this.jquery = 0;
        this.angularJs = 0;
        this.angular = 0;
        this.java = 0;
        this.python = 0;
        this.flask = 0;
        this.aspNetMVC = 0;
        this.aspNetWebForm = 0;
        this.c = 0;
        this.csharp = 0;
        this.nodeJS = 0;
        this.expressNodeJs = 0;
        this.cake = 0;
        this.django = 0;
        this.majento = 0;
        this.php = 0;
        this.vue = 0;
        this.wordpress = 0;
        this.ruby = 0;
        this.mySQLServer = 0;
        this.mySQL = 0;
        this.salesforce = 0;
        this.photoshop = 0;
        this.illustrator = 0;
        this.seo = 0;
        this.laravel = 0;
    }
}